added google search, swap in new databases
