+++ pyproject.toml
@@ -17,6 +17,7 @@ pydantic = "^1.9.0"
 rich = "^12.0.1"
 boto3 = "^1.21.27"
 awscli = "^1.22.82"
+google = "^3.0.0"
 
 [tool.poetry.dev-dependencies]
